ASM compiler: Create a dedicated function for the expression
the ASM compiler creates now a dedicated run-function with all variables
as native double argument and native double return. The run with context
will now delegate to this function. This will improve the performance
when theres more than one lookup for a variable, also this prepares
other improvements in future